+++ CHANGES.rst
@@ -4,7 +4,10 @@ Changelog
 2.0.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+New Features:
+
+- Include translated role title in `@roles` GET.
+  [lgraf]
 
 
 2.0.1 (2018-06-22)